Update vcpkg-scripts-sha, ignore file remove failures when minting the standalone bundles.

+++ vcpkg-init/mint-standalone-bundle.ps1
@@ -75,7 +75,9 @@ try {
         Move-Item 'LICENSE.txt' '../out/LICENSE.txt'
         Move-Item 'triplets' '../out/triplets'
         foreach ($exclusion in $scripts_exclusions) {
-            Remove-Item "scripts/$exclusion" -Recurse -Force
+            if (Test-Path "scripts/$exclusion") {
+                Remove-Item "scripts/$exclusion" -Recurse -Force
+            }
         }
         foreach ($dep in $scripts_dependencies) {
             Move-Item "scripts/$dep" "../out/scripts/$dep"
